+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # pydeps2env
 
+
+## v1.3.0
+
+### added
+
+- support renaming of pip/conda packages based on repository mapping [#57]
+
+### fixed
+
+- remove extras definitions from build system requirements [#57]
+
 ## v1.2.0
 
 ### added